Backend/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
passlib[bcrypt]
spacy
textblob
requests
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz
